Safari-611.1.10/LayoutTests/fast/css/element-scrollByLines-scrollByPages-obsolete.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_equals(document.scrollingElement.scrollByLines, undefined);
    assert_equals(Element.prototype.scrollByLines, undefined);
}, "Element.scrollByLines");

test(function() {
    assert_equals(document.scrollingElement.scrollByPages, undefined);
    assert_equals(Element.prototype.scrollByPages, undefined);
}, "Element.scrollByPages");
</script>
</body>
</html>
